+++ src/main/java/io/github/ladysnake/impersonate/Impersonate.java
@@ -17,33 +17,40 @@
  */
 package io.github.ladysnake.impersonate;
 
+import dev.onyxstudios.cca.api.v3.component.ComponentKey;
+import dev.onyxstudios.cca.api.v3.component.ComponentRegistryV3;
+import dev.onyxstudios.cca.api.v3.entity.EntityComponentFactoryRegistry;
+import dev.onyxstudios.cca.api.v3.entity.EntityComponentInitializer;
 import io.github.ladysnake.impersonate.impl.ImpersonateCommand;
+import io.github.ladysnake.impersonate.impl.ImpersonateGamerules;
 import io.github.ladysnake.impersonate.impl.PlayerImpersonator;
-import nerdhub.cardinal.components.api.ComponentRegistry;
-import nerdhub.cardinal.components.api.ComponentType;
-import nerdhub.cardinal.components.api.event.EntityComponentCallback;
-import nerdhub.cardinal.components.api.util.EntityComponents;
 import nerdhub.cardinal.components.api.util.RespawnCopyStrategy;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.registry.CommandRegistry;
-import net.minecraft.entity.player.PlayerEntity;
+import net.fabricmc.fabric.api.command.v1.CommandRegistrationCallback;
 import net.minecraft.util.Identifier;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 /**
  * Main entrypoint for Impersonate
  *
  * @see Impersonator
- * @see PlayerSkins
  */
-public final class Impersonate implements ModInitializer {
-    public static final ComponentType<Impersonator> IMPERSONATION = ComponentRegistry.INSTANCE.registerIfAbsent(
+public final class Impersonate implements ModInitializer, EntityComponentInitializer {
+    public static final Logger LOGGER = LogManager.getLogger("Impersonate");
+    public static final ComponentKey<Impersonator> IMPERSONATION = ComponentRegistryV3.INSTANCE.getOrCreate(
         new Identifier("impersonate", "impersonation"),
         Impersonator.class
-    ).attach(EntityComponentCallback.event(PlayerEntity.class), PlayerImpersonator::new);
+    );
 
     @Override
     public void onInitialize() {
-        EntityComponents.setRespawnCopyStrategy(IMPERSONATION, RespawnCopyStrategy.ALWAYS_COPY);
-        CommandRegistry.INSTANCE.register(false, ImpersonateCommand::register);
+        CommandRegistrationCallback.EVENT.register((dispatcher, dedicated) -> ImpersonateCommand.register(dispatcher));
+        ImpersonateGamerules.init();
+    }
+
+    @Override
+    public void registerEntityComponentFactories(EntityComponentFactoryRegistry registry) {
+        registry.registerForPlayers(IMPERSONATION, PlayerImpersonator::new, RespawnCopyStrategy.ALWAYS_COPY);
     }
 }