+++ src/main/java/io/github/ladysnake/impersonate/impl/ImpersonateGamerules.java
@@ -17,34 +17,28 @@
  */
 package io.github.ladysnake.impersonate.impl;
 
-import com.mojang.brigadier.arguments.BoolArgumentType;
-import io.github.ladysnake.impersonate.Impersonator;
-import io.github.ladysnake.impersonate.impl.mixin.GameRulesAccessor;
-import io.github.ladysnake.impersonate.impl.mixin.RuleTypeAccessor;
-import net.minecraft.server.MinecraftServer;
+import net.fabricmc.fabric.api.gamerule.v1.GameRuleFactory;
+import net.fabricmc.fabric.api.gamerule.v1.GameRuleRegistry;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.world.GameRules;
 
-import java.util.function.BiConsumer;
-
 public final class ImpersonateGamerules {
-    public static final GameRules.RuleKey<GameRules.BooleanRule> FAKE_CAPES =
-        register("impersonate:fakeCapes", createBooleanRule(false, (server, rule) -> {
+    public static final GameRules.Key<GameRules.BooleanRule> FAKE_CAPES =
+        register("fakeCapes", GameRuleFactory.createBooleanRule(false, (server, rule) -> {
             for (ServerPlayerEntity player : server.getPlayerManager().getPlayerList()) {
-                ((PlayerImpersonator) Impersonator.get(player)).sync();
+                if (rule.get()) {
+                    ((PlayerEntityExtensions)player).impersonate_resetCape();
+                } else {
+                    ((PlayerEntityExtensions) player).impersonate_disableCape();
+                }
             }
         }));
 
-    public static final GameRules.RuleKey<GameRules.BooleanRule> OP_REVEAL_IMPERSONATIONS =
-        register("impersonate:opRevealImpersonations", createBooleanRule(true, (server, rule) -> {}));
+    public static final GameRules.Key<GameRules.BooleanRule> OP_REVEAL_IMPERSONATIONS =
+        register("opRevealImpersonations", GameRuleFactory.createBooleanRule(true));
 
-    @SuppressWarnings("unchecked")
-    private static GameRules.RuleType<GameRules.BooleanRule> createBooleanRule(boolean initialValue, BiConsumer<MinecraftServer, GameRules.BooleanRule> changeCallback) {
-        return RuleTypeAccessor.invokeNew(BoolArgumentType::bool, type -> new GameRules.BooleanRule((GameRules.RuleType<GameRules.BooleanRule>) type, initialValue), changeCallback);
-    }
-
-    private static <T extends GameRules.Rule<T>> GameRules.RuleKey<T> register(String name, GameRules.RuleType<T> type) {
-        return GameRulesAccessor.invokeRegister(name, type);
+    private static <T extends GameRules.Rule<T>> GameRules.Key<T> register(String name, GameRules.Type<T> type) {
+        return GameRuleRegistry.register("impersonate:" + name, GameRules.Category.PLAYER, type);
     }
 
     public static void init() {