+++ src/main/java/io/github/ladysnake/impersonate/impl/ImpersonateText.java
@@ -30,13 +30,17 @@ public class ImpersonateText extends LiteralText implements RecipientAwareText {
     private boolean revealed;
 
     public static Text get(PlayerEntity player) {
-        Impersonator impersonator = Impersonator.get(player);
-        String fakeName = impersonator.getEditedProfile().getName();
-        String trueText = String.format("%s (%s)", fakeName, player.getGameProfile().getName());
-        return new ImpersonateText(trueText, fakeName, false);
+        return get(player, false);
     }
 
-    public ImpersonateText(String trueText, String fakedText, boolean revealed) {
+    public static Text get(PlayerEntity player, boolean reveal) {
+        Impersonator impersonator = Impersonator.get(player);
+        String fakeName = impersonator.getEditedProfile().getName();
+        String trueText = String.format("%s(%s)", fakeName, player.getGameProfile().getName());
+        return new ImpersonateText(trueText, fakeName, reveal);
+    }
+
+    private ImpersonateText(String trueText, String fakedText, boolean revealed) {
         super(fakedText);
         this.trueText = trueText;
         this.fakedText = fakedText;