+++ src/main/java/io/github/ladysnake/impersonate/impl/mixin/EntityTrackerAccessor.java
@@ -17,15 +17,17 @@
  */
 package io.github.ladysnake.impersonate.impl.mixin;
 
-import net.minecraft.world.GameRules;
+import net.minecraft.server.network.EntityTrackerEntry;
+import net.minecraft.server.network.ServerPlayerEntity;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Invoker;
+import org.spongepowered.asm.mixin.gen.Accessor;
 
-@Mixin(GameRules.class)
-public interface GameRulesAccessor {
-    @Invoker
-    @SuppressWarnings("PublicStaticMixinMember")
-    static <T extends GameRules.Rule<T>> GameRules.RuleKey<T> invokeRegister(String name, GameRules.RuleType<T> type) {
-        throw new AssertionError();
-    }
+import java.util.Set;
+
+@Mixin(targets = "net.minecraft.server.world.ThreadedAnvilChunkStorage$EntityTracker")
+public interface EntityTrackerAccessor {
+    @Accessor
+    EntityTrackerEntry getEntry();
+    @Accessor
+    Set<ServerPlayerEntity> getPlayersTracking();
 }