+++ src/main/java/io/github/ladysnake/impersonate/impl/mixin/GameMessageS2CPacketAccessor.java
@@ -15,12 +15,18 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; If not, see <https://www.gnu.org/licenses>.
  */
-package io.github.ladysnake.impersonate.impl;
+package io.github.ladysnake.impersonate.impl.mixin;
 
-import com.mojang.authlib.GameProfile;
+import net.minecraft.network.packet.s2c.play.GameMessageS2CPacket;
+import net.minecraft.text.Text;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
 
-public interface PlayerEntityExtensions {
-    GameProfile impersonate_getActualGameProfile();
-    void impersonate_resetCape();
-    void impersonate_disableCape();
+@Mixin(GameMessageS2CPacket.class)
+public interface GameMessageS2CPacketAccessor {
+    @Accessor
+    Text getMessage();
+
+    @Accessor
+    void setMessage(Text message);
 }