+++ src/main/java/io/github/ladysnake/impersonate/impl/mixin/PlayerEntityMixin.java
@@ -18,12 +18,13 @@
 package io.github.ladysnake.impersonate.impl.mixin;
 
 import com.mojang.authlib.GameProfile;
-import io.github.ladysnake.impersonate.Impersonate;
 import io.github.ladysnake.impersonate.Impersonator;
 import io.github.ladysnake.impersonate.impl.ImpersonateText;
 import io.github.ladysnake.impersonate.impl.PlayerEntityExtensions;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.data.DataTracker;
+import net.minecraft.entity.data.TrackedData;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.text.Text;
 import net.minecraft.world.World;
@@ -33,7 +34,6 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(PlayerEntity.class)
@@ -42,16 +42,15 @@ public abstract class PlayerEntityMixin extends LivingEntity implements PlayerEn
     @Final
     private GameProfile gameProfile;
 
+    @Shadow
+    @Final
+    protected static TrackedData<Byte> PLAYER_MODEL_PARTS;
+
     @Unique
-    protected Impersonator impersonate_self = Impersonate.IMPERSONATION.get(this);    // cache the component for faster access
+    private boolean wantsCapeDisplay;
 
     protected PlayerEntityMixin(EntityType<? extends LivingEntity> type, World world) {
         super(type, world);
-    }
-
-    @Override
-    public Impersonator impersonate_getAsImpersonator() {
-        return impersonate_self;
     }
 
     @Override
@@ -59,40 +58,31 @@ public abstract class PlayerEntityMixin extends LivingEntity implements PlayerEn
         return this.gameProfile;
     }
 
-    /**
-     * Fakes the player's game profile on clients
-     */
-    @Inject(method = "getGameProfile", at = @At("HEAD"), cancellable = true)
-    private void fakeGameProfile(CallbackInfoReturnable<GameProfile> cir) {
-        if (this.world.isClient) {
-            if (impersonate_self.isImpersonating()) {
-                cir.setReturnValue(impersonate_self.getEditedProfile());
-            }
+    @Override
+    public void impersonate_resetCape() {
+        if (this.wantsCapeDisplay) {
+            DataTracker dataTracker = this.getDataTracker();
+            byte modelMask = dataTracker.get(PLAYER_MODEL_PARTS);
+            byte newModelMask = (byte) (modelMask | 1);
+            dataTracker.set(PLAYER_MODEL_PARTS, newModelMask);
         }
     }
 
-    @ModifyArg(method = "getDisplayName", at = @At(value = "INVOKE", target = "Lnet/minecraft/scoreboard/Team;modifyText(Lnet/minecraft/scoreboard/AbstractTeam;Lnet/minecraft/text/Text;)Lnet/minecraft/text/Text;"))
-    private Text fakeDisplayName(Text original) {
-        // No need to fake on clients, as #fakeGameProfile already covers it
-        if (!world.isClient && impersonate_self.isImpersonating()) {
-            return ImpersonateText.get((PlayerEntity) (Object) this);
-        }
-        return original;
+    @Override
+    public void impersonate_disableCape() {
+        DataTracker dataTracker = this.getDataTracker();
+        byte modelMask = dataTracker.get(PLAYER_MODEL_PARTS);
+        this.wantsCapeDisplay = (modelMask & 1) != 0;
+        byte newModelMask = (byte) (modelMask & ~1);
+        dataTracker.set(PLAYER_MODEL_PARTS, newModelMask);
     }
 
-    @ModifyArg(method = "getNameAndUuid", at = @At(value = "INVOKE", target = "Lnet/minecraft/text/LiteralText;append(Lnet/minecraft/text/Text;)Lnet/minecraft/text/Text;", ordinal = 0))
-    private Text fakeNameAndUuid(Text originalName) {
-        if (impersonate_self.isImpersonating()) {
-            return ImpersonateText.get((PlayerEntity) (Object) this);
+    @Inject(method = "getName", at = @At("RETURN"), cancellable = true)
+    private void fakeName(CallbackInfoReturnable<Text> cir) {
+        PlayerEntity self = ((PlayerEntity) (Object) this);
+        if (Impersonator.get(self).isImpersonating()) {
+            // if the client is aware that there is an impersonation, they should display it
+            cir.setReturnValue(ImpersonateText.get(self, world.isClient));
         }
-        return originalName;
-    }
-
-    @ModifyArg(method = "getNameAndUuid", at = @At(value = "INVOKE", target = "Lnet/minecraft/text/Text;append(Ljava/lang/String;)Lnet/minecraft/text/Text;", ordinal = 1))
-    private String fakeNameAndUuid(String originalUuid) {
-        if (impersonate_self.isImpersonating()) {
-            return impersonate_getActualGameProfile().getId().toString();
-        }
-        return originalUuid;
     }
 }