+++ src/main/java/io/github/ladysnake/impersonate/impl/mixin/PlayerListS2CPacketEntryAccessor.java
@@ -17,20 +17,16 @@
  */
 package io.github.ladysnake.impersonate.impl.mixin;
 
-import com.mojang.brigadier.arguments.ArgumentType;
-import net.minecraft.world.GameRules;
+import com.mojang.authlib.GameProfile;
+import net.minecraft.network.packet.s2c.play.PlayerListS2CPacket;
+import net.minecraft.text.Text;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Invoker;
+import org.spongepowered.asm.mixin.gen.Accessor;
 
-import java.util.function.BiConsumer;
-import java.util.function.Function;
-import java.util.function.Supplier;
-
-@Mixin(GameRules.RuleType.class)
-public interface RuleTypeAccessor {
-    @SuppressWarnings({"PublicStaticMixinMember", "rawtypes"})  // Mixin's AP shits the bed with recursive generics (see issue #384)
-    @Invoker("<init>")
-    static GameRules.RuleType invokeNew(Supplier<ArgumentType<?>> argumentType, Function factory, BiConsumer notifier) {
-        throw new AssertionError();
-    }
+@Mixin(PlayerListS2CPacket.Entry.class)
+public interface PlayerListS2CPacketEntryAccessor {
+    @Accessor
+    void setDisplayName(Text displayName);
+    @Accessor
+    void setProfile(GameProfile profile);
 }