+++ src/main/java/io/github/ladysnake/impersonate/impl/mixin/ServerPlayNetworkHandlerMixin.java
@@ -17,28 +17,54 @@
  */
 package io.github.ladysnake.impersonate.impl.mixin;
 
-import io.github.ladysnake.impersonate.impl.RecipientAwareText;
-import io.netty.util.concurrent.Future;
-import io.netty.util.concurrent.GenericFutureListener;
+import io.github.ladysnake.impersonate.Impersonator;
+import io.github.ladysnake.impersonate.impl.PacketMeddling;
 import net.minecraft.network.Packet;
+import net.minecraft.network.packet.s2c.play.GameMessageS2CPacket;
+import net.minecraft.network.packet.s2c.play.PlayerListS2CPacket;
 import net.minecraft.server.network.ServerPlayNetworkHandler;
 import net.minecraft.server.network.ServerPlayerEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
 
 @Mixin(ServerPlayNetworkHandler.class)
 public abstract class ServerPlayNetworkHandlerMixin {
     @Shadow
     public ServerPlayerEntity player;
 
-    @Inject(method = "sendPacket(Lnet/minecraft/network/Packet;Lio/netty/util/concurrent/GenericFutureListener;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/network/ClientConnection;send(Lnet/minecraft/network/Packet;Lio/netty/util/concurrent/GenericFutureListener;)V"))
-    private void resolveFakeTextsInPackets(Packet<?> packet, GenericFutureListener<? extends Future<? super Void>> genericFutureListener, CallbackInfo ci) {
-        if (packet instanceof ChatMessageS2CPacketAccessor) {
-            ChatMessageS2CPacketAccessor accessiblePacket = (ChatMessageS2CPacketAccessor) packet;
-            accessiblePacket.setMessage(((RecipientAwareText) accessiblePacket.getMessage()).impersonateResolveAll(this.player));
+    @ModifyArg(method = "sendPacket(Lnet/minecraft/network/Packet;Lio/netty/util/concurrent/GenericFutureListener;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/network/ClientConnection;send(Lnet/minecraft/network/Packet;Lio/netty/util/concurrent/GenericFutureListener;)V"))
+    private Packet<?> resolveFakeTextsInPackets(Packet<?> packet) {
+        if (packet instanceof GameMessageS2CPacket) {
+            if (this.existsImpersonator()) {
+                // copying the packet converts hidden texts to literals, so we must resolve it before that
+                PacketMeddling.resolveFakeTexts((GameMessageS2CPacket) packet, player);
+                GameMessageS2CPacket copy = new GameMessageS2CPacket();
+                if (PacketMeddling.copyPacket(packet, copy)) {
+                    return copy;
+                }
+            }
+        } else if (packet instanceof PlayerListS2CPacket) {
+            if (this.existsImpersonator()) {
+                PlayerListS2CPacket copy = new PlayerListS2CPacket();
+                if (PacketMeddling.copyPacket(packet, copy)) {
+                    PacketMeddling.resolvePlayerListEntries(copy, player);
+                    return copy;
+                }
+            }
         }
+        return packet;
+    }
+
+    @Unique
+    private boolean existsImpersonator() {
+        for (ServerPlayerEntity player : this.player.server.getPlayerManager().getPlayerList()) {
+            if (Impersonator.get(player).isImpersonating()) {
+                return true;
+            }
+        }
+        return false;
     }
 }