+++ src/main/java/io/github/ladysnake/impersonate/impl/mixin/ServerPlayerEntityMixin.java
@@ -17,18 +17,30 @@
  */
 package io.github.ladysnake.impersonate.impl.mixin;
 
+import com.mojang.authlib.GameProfile;
 import io.github.ladysnake.impersonate.impl.ImpersonateGamerules;
-import net.minecraft.world.GameRules;
+import io.github.ladysnake.impersonate.impl.PlayerEntityExtensions;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.network.packet.c2s.play.ClientSettingsC2SPacket;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(GameRules.class)
-public abstract class GameRulesMixin {
-    @SuppressWarnings("UnresolvedMixinReference")
-    @Inject(method = "<clinit>", at = @At("RETURN"))
-    private static void initRequiemGamerules(CallbackInfo ci) {
-        ImpersonateGamerules.init();
+@Mixin(ServerPlayerEntity.class)
+public abstract class ServerPlayerEntityMixin extends PlayerEntity implements PlayerEntityExtensions {
+
+    public ServerPlayerEntityMixin(World world, BlockPos pos, float yaw, GameProfile profile) {
+        super(world, pos, yaw, profile);
+    }
+
+    @Inject(method = "setClientSettings", at = @At("RETURN"))
+    private void removeCapeIfDisallowed(ClientSettingsC2SPacket clientSettingsC2SPacket, CallbackInfo ci) {
+        if (this.world.getGameRules().getBoolean(ImpersonateGamerules.FAKE_CAPES)) {
+            this.impersonate_disableCape();
+        }
     }
 }