+++ src/main/java/io/github/ladysnake/impersonate/impl/mixin/TranslatableTextMixin.java
@@ -17,16 +17,29 @@
  */
 package io.github.ladysnake.impersonate.impl.mixin;
 
-import net.minecraft.client.network.packet.ChatMessageS2CPacket;
-import net.minecraft.text.Text;
+import io.github.ladysnake.impersonate.impl.RecipientAwareText;
+import net.minecraft.server.command.CommandOutput;
+import net.minecraft.text.TranslatableText;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
+import org.spongepowered.asm.mixin.Shadow;
 
-@Mixin(ChatMessageS2CPacket.class)
-public interface ChatMessageS2CPacketAccessor {
-    @Accessor
-    Text getMessage();
+@Mixin(TranslatableText.class)
+public abstract class TranslatableTextMixin implements RecipientAwareText {
+    @Shadow
+    @Final
+    private Object[] args;
 
-    @Accessor
-    void setMessage(Text message);
+    @Shadow
+    public abstract Object[] getArgs();
+
+    @Override
+    public void impersonateResolve(CommandOutput recipient) {
+        for (int i = 0; i < this.getArgs().length; i++) {
+            Object arg = args[i];
+            if (arg instanceof RecipientAwareText) {
+                args[i] = ((RecipientAwareText) arg).impersonateResolveAll(recipient);
+            }
+        }
+    }
 }